refactor: split locations routes and decompose AdminUserDetailPage
locations.ts (686 lines) into routes/locations/{crud,invite,members,
stats,index}, matching the routes/<name>/{...,index}.ts pattern. Invite
code generation extracted to lib/inviteCode.ts so non-route callers
(admin reset flows) don't reach into a route file.

AdminUserDetailPage.tsx into a thin orchestrator over focused cards
(Identity, Stats, Credits, Security, Overrides, Actions) plus
Dialogs/Skeleton/StatItems leaves and useAdminUserActions /
useAdminUserOverrides hooks. Constants hoisted to
adminUserDetailConstants.ts.
