+++ server/src/lib/inviteCode.ts
@@ -1,0 +1,5 @@
+import crypto from 'node:crypto';
+
+export function generateInviteCode(): string {
+  return crypto.randomBytes(16).toString('hex');
+}
