+++ src/features/admin/AdminUserSkeleton.tsx
@@ -1,0 +1,39 @@
+import { Card, CardContent, CardHeader } from '@/components/ui/card';
+import { PageHeader } from '@/components/ui/page-header';
+import { Skeleton } from '@/components/ui/skeleton';
+
+export function AdminUserSkeleton() {
+  return (
+    <div className="page-content">
+      <PageHeader title="" back />
+      <Card>
+        <CardHeader><Skeleton className="h-5 w-20" /></CardHeader>
+        <CardContent>
+          <div className="grid grid-cols-1 sm:grid-cols-2 gap-3">
+            {Array.from({ length: 6 }, (_, i) => (
+              // biome-ignore lint/suspicious/noArrayIndexKey: skeleton placeholders have no stable identity
+              <div key={i} className="space-y-1">
+                <Skeleton className="h-3 w-16" />
+                <Skeleton className="h-4 w-32" />
+              </div>
+            ))}
+          </div>
+        </CardContent>
+      </Card>
+      <Card>
+        <CardHeader><Skeleton className="h-5 w-14" /></CardHeader>
+        <CardContent>
+          <div className="grid grid-cols-2 sm:grid-cols-3 lg:grid-cols-6 gap-2">
+            {Array.from({ length: 6 }, (_, i) => (
+              // biome-ignore lint/suspicious/noArrayIndexKey: skeleton placeholders have no stable identity
+              <div key={i} className="p-3 rounded-[var(--radius-sm)] bg-[var(--bg-input)]">
+                <Skeleton className="h-3 w-14 mb-1" />
+                <Skeleton className="h-5 w-8" />
+              </div>
+            ))}
+          </div>
+        </CardContent>
+      </Card>
+    </div>
+  );
+}
